+++ utils/metrics.py
@@ -51,9 +51,9 @@ class CityEvalMIoU(tf.keras.metrics.MeanIoU):
         y_true = tf.squeeze(y_true, axis=-1)
         y_pred += 1
 
-        zeros_y_pred = tf.zeros(tf.shape(y_pred), tf.int64)
+        zeros_y_pred = tf.zeros(tf.shape(y_pred), tf.int32)
         zeros_y_pred += y_pred
-        indices = tf.cast(tf.where(tf.equal(y_true, 0), 0, 1), tf.int64)
+        indices = tf.cast(tf.where(tf.equal(y_true, 0), 0, 1), tf.int32)
 
         y_true *= indices
         zeros_y_pred *= indices
